requirements.txt
mysql-connector-python==8.0.16
Pillow==6.1.0
pytesseract==0.3.0
selenium==3.141.0
SQLAlchemy==1.3.7
urllib3==1.25.3
